// projects/adminmodule/src/app/models/responsemodel.ts
export interface Ires_Webportal {
    Batch_code: number
    Batch_name: string
    Startdatetime: string
    Enddatetime: string
    Createdby: number
    Editedby: number
}